+++ grails-app/init/io/xh/hoist/BootStrap.groovy
@@ -101,15 +101,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhActivityTrackingConfig',
                 valueType: 'json',
-                defaultValue: [
-                    clientHealthReport: [intervalMins: -1],
-                    enabled: true,
-                    levels: [[username: '*', category: '*', severity: 'INFO']],
-                    logData: false,
-                    maxDataLength: 2000,
-                    maxEntriesPerMin: 1000,
-                    maxRows: [default: 10000, limit: 25000, options: [1000, 5000, 10000, 25000]]
-                ],
+                defaultValue: [:],
                 typedClass: ActivityTrackingConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
@@ -118,7 +110,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhAlertBannerConfig',
                 valueType: 'json',
-                defaultValue: [enabled: true],
+                defaultValue: [:],
                 typedClass: AlertBannerConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
@@ -151,7 +143,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhChangelogConfig',
                 valueType: 'json',
-                defaultValue: [enabled: true, excludedVersions: [], excludedCategories: [], limitToRoles: []],
+                defaultValue: [:],
                 typedClass: ChangelogConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
@@ -160,7 +152,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhClientErrorConfig',
                 valueType: 'json',
-                defaultValue: [intervalMins: 2],
+                defaultValue: [:],
                 typedClass: ClientErrorConfig,
                 groupName: 'xh.io',
                 note: 'Configures handling of client error reports. Errors are queued when received and processed every [intervalMins].'
@@ -168,12 +160,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhConnPoolMonitoringConfig',
                 valueType: 'json',
-                defaultValue: [
-                    enabled: true,
-                    snapshotInterval: 60,
-                    maxSnapshots: 1440,
-                    writeToLog: false
-                ],
+                defaultValue: [:],
                 typedClass: ConnPoolMonitoringConfig,
                 groupName: 'xh.io',
                 note: 'Configures built-in JDBC connection pool monitoring.'
@@ -241,10 +228,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhEnvPollConfig',
                 valueType: 'json',
-                defaultValue: [
-                    interval: 10,
-                    onVersionChange: configService.getMap('xhAppVersionCheck', [mode: 'promptReload']).get('mode')
-                ],
+                defaultValue: [:],
                 typedClass: EnvPollConfig,
                 groupName: 'xh.io',
                 note: "Controls client calls to server to poll for version, instance changes, or auth changes. Supports the following options:\n\n" +
@@ -264,10 +248,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhExportConfig',
                 valueType: 'json',
-                defaultValue: [
-                    streamingCellThreshold: 100000,
-                    toastCellThreshold: 3000
-                ],
+                defaultValue: [:],
                 typedClass: ExportConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
@@ -284,7 +265,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhIdleConfig',
                 valueType: 'json',
-                defaultValue: [timeout: 120, appTimeouts: [:]],
+                defaultValue: [:],
                 typedClass: IdleConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
@@ -293,20 +274,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhLdapConfig',
                 valueType: 'json',
-                defaultValue: [
-                    enabled: false,
-                    timeoutMs: 60000,
-                    cacheExpireSecs: 300,
-                    useMatchingRuleInChain: false,
-                    skipTlsCertVerification: false,
-                    servers: [
-                        [
-                            host: '',
-                            baseUserDn: '',
-                            baseGroupDn: '',
-                        ]
-                    ]
-                ],
+                defaultValue: [:],
                 typedClass: LdapConfig,
                 groupName: 'xh.io',
                 note: 'Supports connecting to LDAP servers.'
@@ -326,10 +294,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhLogArchiveConfig',
                 valueType: 'json',
-                defaultValue: [
-                    archiveAfterDays: 30,
-                    archiveFolder: 'archive'
-                ],
+                defaultValue: [:],
                 typedClass: LogArchiveConfig,
                 groupName: 'xh.io',
                 note: 'Configures automatic cleanup and archiving of log files. Files older than "archiveAfterDays" will be moved into zipped bundles within the specified "archiveFolder".'
@@ -337,15 +302,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhMemoryMonitoringConfig',
                 valueType: 'json',
-                defaultValue: [
-                    enabled: true,
-                    snapshotInterval: 60,
-                    maxSnapshots: 1440,
-                    heapDumpDir: null,
-                    preservePastInstances: true,
-                    maxPastInstances: 10,
-                    writeToLog: true
-                ],
+                defaultValue: [:],
                 typedClass: MemoryMonitoringConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
@@ -354,15 +311,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhMonitorConfig',
                 valueType: 'json',
-                defaultValue: [
-                    monitorRefreshMins: 10,
-                    failNotifyThreshold: 2,
-                    warnNotifyThreshold: 5,
-                    monitorStartupDelayMins: 1,
-                    monitorRepeatNotifyMins: 60,
-                    monitorTimeoutSecs: 15,
-                    writeToMonitorLog: true
-                ],
+                defaultValue: [:],
                 typedClass: MonitorConfig,
                 groupName: 'xh.io',
                 note: 'Configures server-side status monitoring and notifications. Note failNotifyThreshold and warnNotifyThreshold are the number of refresh cycles a monitor will need to be in said status to trigger "alertMode".'
@@ -377,12 +326,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhMetricsConfig',
                 valueType: 'json',
-                defaultValue: [
-                    prometheusEnabled: false,
-                    otlpEnabled: false,
-                    prometheusConfig: [:],
-                    otlpConfig: [:]
-                ],
+                defaultValue: [:],
                 typedClass: MetricsConfig,
                 groupName: 'xh.io',
                 note: 'Parameters for observable metric support'
@@ -390,14 +334,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhTraceConfig',
                 valueType: 'json',
-                defaultValue: [
-                    enabled: false,
-                    sampleRate: 1.0,
-                    sampleRules: [],
-                    jdbcTracingEnabled: false,
-                    otlpEnabled: false,
-                    otlpConfig: [:]
-                ],
+                defaultValue: [:],
                 typedClass: TraceConfig,
                 clientVisible: true,
                 groupName: 'xh.io',
@@ -413,10 +350,7 @@ class BootStrap implements LogSupport {
             new ConfigSpec(
                 name: 'xhWebSocketConfig',
                 valueType: 'json',
-                defaultValue: [
-                    sendTimeLimitMs: 1000,
-                    bufferSizeLimitBytes: 1000000
-                ],
+                defaultValue: [:],
                 typedClass: WebSocketConfig,
                 groupName: 'xh.io',
                 note: 'Parameters for the managed WebSocket sessions created by Hoist.'